Added ReminderProcessor for those with UserID only and TaskProcessor compilation support for null UserID.

+++ TeamSupport.ServiceLibrary/ReminderProcessor.cs
@@ -42,7 +42,7 @@ namespace TeamSupport.ServiceLibrary
       Logs.WriteEvent("***********************************************************************************");
       
       MailMessage message;
-      UsersViewItem user = UsersView.GetUsersViewItem(LoginUser, reminder.UserID);
+      UsersViewItem user = UsersView.GetUsersViewItem(LoginUser, (int)reminder.UserID);
       if (user == null) return;
       string description = "";
       switch (reminder.RefType)
@@ -95,7 +95,7 @@ namespace TeamSupport.ServiceLibrary
           description = String.Format("Reminder sent to {0} for Task {1}", message.To.ToString(), task.TaskName);
           Logs.WriteEvent(description);
           ActionLogs.AddActionLog(LoginUser, ActionLogType.Insert, ReferenceType.Tasks, reminder.ReminderID, description);
-          ActionLogs.AddActionLog(LoginUser, ActionLogType.Insert, ReferenceType.Users, reminder.UserID, description);
+          ActionLogs.AddActionLog(LoginUser, ActionLogType.Insert, ReferenceType.Users, (int)reminder.UserID, description);
           break;
         default:
           message = null;
